+++ kadepy/__init__.py
@@ -1,26 +1,26 @@
-from ._core import (
-    create_swarm,
-    xor_distance,
-    set_callback,
-    dht_ping,
-    dht_find_node,
-    dht_announce_peer,
-    dht_get_peers,
-    dump_routing_table
-)
+# from ._core import (
+#     create_swarm,
+#     xor_distance,
+#     set_callback,
+#     dht_ping,
+#     dht_find_node,
+#     dht_announce_peer,
+#     dht_get_peers,
+#     dump_routing_table
+# )
 
 from .swarm import Swarm
 
-__version__ = "0.2.4"
+__version__ = "0.3.0"
 
 __all__ = [
-    'create_swarm',
-    'xor_distance',
-    'set_callback',
-    'dht_ping',
-    'dht_find_node',
-    'dht_announce_peer',
-    'dht_get_peers',
-    'dump_routing_table',
+    # 'create_swarm',
+    # 'xor_distance',
+    # 'set_callback',
+    # 'dht_ping',
+    # 'dht_find_node',
+    # 'dht_announce_peer',
+    # 'dht_get_peers',
+    # 'dump_routing_table',
     'Swarm'
 ]